requirement.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
Django==3.1.2
django-appconf==1.0.4
django-environ==0.4.5
django-imagekit==4.0.2
filelock==3.0.12
pilkit==2.0
Pillow==8.0.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.0.32
